requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain - Compatible versions
langchain==0.1.16
langchain-community>=0.0.32
langchain-core==0.1.53
langsmith>=0.1.17,<0.2.0
langchain-groq==0.0.1
langchain-huggingface==0.0.1

# Embeddings & Transformers
sentence-transformers>=2.6.0
transformers>=4.39.0
tokenizers>=0.19.1
huggingface-hub>=0.23.0

# Vector Databases
chromadb==0.4.22
faiss-cpu==1.7.4

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Web Search & API
tavily-python==0.3.3
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.3

# Database
sqlalchemy==2.0.25

# Environment & File Handling
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional Dependencies
numpy<2,>=1
tenacity<9.0.0,>=8.1.0
PyYAML>=5.3
dataclasses-json<0.7,>=0.5.7
jsonpatch<2.0,>=1.33
typing-extensions>=4.8.0
anyio<4.0.0,>=3.7.1
click>=7.0
certifi>=2017.4.17
charset-normalizer<4,>=2
idna<4,>=2.5
urllib3<3,>=1.21.1